vio2sf: optimize. Simplify SPU logic. Use single-precision floating point type.

+++ 3rdparty/vio2sf/desmume/resampler.h
@@ -52,7 +52,7 @@ void resampler_set_quality(void *, int quality);
 int resampler_get_free_count(void *);
 void resampler_write_sample(void *, short sample);
 void resampler_write_sample_fixed(void *, int sample, unsigned char depth);
-void resampler_set_rate( void *, double new_factor );
+void resampler_set_rate( void *, float new_factor );
 int resampler_ready(void *);
 void resampler_clear(void *);
 int resampler_get_sample_count(void *);
